refactor(commands): extract /managed-agents command into managed_agents.rs
